+++ Instagram/src/Home/index.js
@@ -1,11 +1,11 @@
 import { LinearGradient } from "expo-linear-gradient";
 import React from "react";
-import { Text, View, StyleSheet, Image } from "react-native";
+import styles from './styles'
+import { Text, View, Image } from "react-native";
 
 const HomePage = ({ }) => {
   return (
     <View style={styles.container}>
-      {/* Cabeçalho / Notificações / Direct Messenger */}
       <View style={styles.header}>
         <Image source={require("../../assets/instagram-text-icon.png")} style={styles.logoHeader} />
         <View style={styles.headerNav}>
@@ -14,7 +14,6 @@ const HomePage = ({ }) => {
         </View>
       </View>
 
-      {/* Story User */}
       <View style={styles.containerStories}>
         <View style={styles.containerStory}>
           <View style={styles.containerStoryAdd}>
@@ -24,7 +23,6 @@ const HomePage = ({ }) => {
           <Text>Your story</Text>
         </View>
 
-        {/* Stories geral*/}
         <View style={styles.containerStory}>
           <LinearGradient
             colors={["#F58529", "#DD2A7B", "#8134AF"]}
@@ -47,7 +45,7 @@ const HomePage = ({ }) => {
         </View>
         <View style={styles.containerStory}>
           <LinearGradient
-            colors={["#F58529", "#DD2A7B", "#8134AF"]}
+            colors={["#F58528", "#DD2A7D", "#8134AJ"]}
             start={{ x: 0, y: 0.8 }}
             end={{ x: 0.4, y: 0 }}
             style={styles.gradient}
@@ -57,10 +55,8 @@ const HomePage = ({ }) => {
         </View>
       </View>
 
-      {/* Feed */}
       <View>
 
-        {/* Cabeçalho do Feed */}
         <View style={styles.containerFeedHeader}>
           <View style={styles.userContainer}>
             <Image source={require("../../assets/profile.jpeg")} style={styles.userFeed} />
@@ -69,10 +65,8 @@ const HomePage = ({ }) => {
           <Image source={require("../../assets/more.png")} style={styles.more} />
         </View>
 
-        {/* Imagem do Feed */}
         <Image source={require("../../assets/profile.jpeg")} style={styles.feed} />
 
-        {/* Rodapé do Feed */}
         <View style={styles.containerFeedFooter}>
           <View style={styles.FooterAction}>
             <Image source={require("../../assets/curtida-icone.png")} style={styles.iconMenu} />
@@ -107,152 +101,5 @@ const HomePage = ({ }) => {
     </View>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: "#fff",
-    gap: 12,
-  },
-  // Cabeçalho / Notificações / Direct Messenger
-  header: {
-    flex: 1,
-    flexDirection: "row",
-    maxHeight: 56,
-    paddingVertical: 10,
-    paddingHorizontal: 20,
-    marginTop: 44,
-  },
-  logoHeader: {
-    objectFit: "contain",
-    width: 124,
-    height: 50,
-  },
-  headerNav: {
-    flex: 1,
-    flexDirection: "row",
-    gap: 20,
-    alignItems: "center",
-    justifyContent: "flex-end",
-  },
-  iconMenu: {
-    objectFit: "contain",
-    width: 24,
-    height: 24,
-  },
-  // Stories
-  containerStories: {
-    height: 102,
-    flexDirection: "row",
-    alignItems: "center",
-    gap: 8,
-    paddingLeft: 4,
-  },
-  containerStory: {
-    padding: 3.5,
-    alignItems: "center",
-    gap: 5,
-  },
-  gradient: {
-    position: "absolute",
-    left: 0,
-    right: 0,
-    height: "100%",
-    borderRadius: 50,
-  },
-  story: {
-    objectFit: "contain",
-    width: 88,
-    height: 88,
-    borderRadius: 50,
-    border: "solid",
-    borderColor: "#fff",
-    borderWidth: 2,
-  },
-  containerStoryAdd: {
-    position: "absolute",
-    right: 4,
-    bottom: 24,
-    zIndex: 1,
-    width: 28,
-    height: 28,
-    borderRadius: 50,
-    border: "solid",
-    borderColor: "#fff",
-    borderWidth: 2,
-    backgroundColor: "#111",
-    justifyContent: "center",
-    alignItems: "center",
-  },
-  storyAdd: {
-    objectFit: "contain",
-    width: 18,
-    height: 18,
-  },
-  // Feed
-  containerFeedHeader: {
-    flexDirection: "row",
-    alignItems: "center",
-    justifyContent: "space-between",
-    gap: 8,
-    padding: 8,
-  },
-  userContainer: {
-    flexDirection: "row",
-    gap: 8,
-    alignItems: "center",
-  },
-  userFeed: {
-    objectFit: "contain",
-    width: 36,
-    height: 36,
-    borderRadius: 50,
-  },
-  userName: {
-    fontSize: 14,
-    fontWeight: "bold",
-  },
-  more: {
-    objectFit: "contain",
-    width: 28,
-    height: 28,
-  },
-  feed: {
-    width: "100%",
-    height: 400,
-    objectFit: "cover",
-  },
-  containerFeedFooter: {
-    flexDirection: "row",
-    alignItems: "center",
-    justifyContent: "space-between",
-    padding: 8,
-  },
-  FooterAction: {
-    flexDirection: "row",
-    gap: 16
-  },
-  PostSave: {
-    marginRight: 8
-  },
-  PostInfo: {
-    marginTop: 4,
-    paddingHorizontal: 8,
-  },
-  PostText: {
-    fontSize: 16,
-  },
-  // Action Bar
-  containerActionBar: {
-    flexDirection: "row",
-    justifyContent: "space-between",
-    alignItems: "center",
-    paddingVertical: 8,
-    paddingHorizontal: 16,
-    marginTop: 28,
-    borderTopWidth: 1,
-    borderTopColor: "#ddd",
-  },
-});
 
 export default HomePage
